more refacto
